Title: 진법 변환 4, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231114/진법 변환 4/base-conversion-4.c
@@ -3,10 +3,10 @@
 int main() {
     long long n;
     int r = 0;
-    int m = 1;
+    long long m = 1;
     scanf("%lld", &n);
     while(n > 0) {
-        long long temp = n%10;
+        int temp = n%10;
         r += temp*m;
         m*=2;
         n/=10;
